Unify the two copies of the settings slider
