+++ web/src/components/DetailPanel.tsx
@@ -8,6 +8,7 @@ import { NameField } from "./NameField";
 import { PluginNodeView, usePluginSurface } from "./PluginNode";
 import { RiskGauge } from "./RiskGauge";
 import { SaveStatus } from "./SaveStatus";
+import { Slider } from "./Slider";
 import { Sparkline } from "./Sparkline";
 import { Toggle } from "./Toggle";
 
@@ -17,46 +18,6 @@ function Section({ title, children }: { title: string; children: React.ReactNode
       <h3 className="display text-[0.68rem] font-semibold tracking-[0.24em] text-text-2 mb-3">{title}</h3>
       {children}
     </section>
-  );
-}
-
-function Slider({
-  label,
-  value,
-  min,
-  max,
-  step,
-  hint,
-  format = (v) => v.toFixed(2),
-  onChange,
-}: {
-  label: string;
-  value: number;
-  min: number;
-  max: number;
-  step: number;
-  hint?: string;
-  format?: (v: number) => string;
-  onChange: (v: number) => void;
-}) {
-  const hintId = useId();
-  return (
-    <label className="block">
-      <div className="flex justify-between mb-1">
-        <span className="label">{label}</span>
-        <span className="mono text-[0.72rem] text-text-0">{format(value)}</span>
-      </div>
-      <input
-        type="range"
-        min={min}
-        max={max}
-        step={step}
-        value={value}
-        aria-describedby={hint ? hintId : undefined}
-        onChange={(e) => onChange(Number(e.target.value))}
-      />
-      {hint && <p id={hintId} className="text-[0.7rem] leading-snug text-text-2 mt-1">{hint}</p>}
-    </label>
   );
 }
